# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIX for JSSR",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/customer": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Get All Customer",
                "description": "Get all customer data",
                "operationId": "customerGetAll",
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Create Customer",
                "description": "Create customer",
                "operationId": "customerCreate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/customer/{id}": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Get Customer",
                "description": "Get customer data",
                "operationId": "customerGet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Customer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Customer"
                ],
                "summary": "Update Customer",
                "description": "Update customer",
                "operationId": "customerUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Customer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}